fix: refactor createSleepPlan function for testing
