+++ web/src/util/utils.ts
@@ -1,3 +1,6 @@
+import { User } from '@auth0/auth0-react';
+import dayjs from 'dayjs';
+
 export function openSidebar() {
   if (typeof window !== 'undefined') {
     document.body.style.overflow = 'hidden';
@@ -57,3 +60,70 @@ export function formatTimeTo12HourFormat(dateString: string | null) {
   const formattedTime = `${formattedHours}:${formattedMinutes} ${period}`;
   return formattedTime;
 }
+
+export const createSleepPlan = (
+  wakeUpTime: dayjs.Dayjs | null,
+  earliestGetReadyTime: dayjs.Dayjs | null,
+  desiredBedTime: dayjs.Dayjs | null,
+  wakeWindows: number[],
+  numOfNaps: number,
+  userId: string | undefined,
+  user: string | User | undefined
+) => {
+  // Create reminders
+  const morningRise = [
+    {
+      description: 'Morning Rise',
+      startTime: wakeUpTime?.toDate()
+    }
+  ];
+
+  const naps: {
+    description: string;
+    startTime: Date | null;
+    endTime: Date | null;
+  }[] = [];
+
+  // Compute how long each nap is
+  const totalDayTime = desiredBedTime?.diff(wakeUpTime, 'hour') ?? 12;
+  const totalWakeWindows = wakeWindows.reduce((accumulator, currentValue) => {
+    return accumulator + currentValue;
+  }, 0);
+  const napTime = (totalDayTime - totalWakeWindows) / numOfNaps;
+
+  // Create naps based on wakeUpTime, wakeWindowss, and desiredBedTime
+  let napEndTime = wakeUpTime;
+  for (let index = 0; index < wakeWindows.length - 1; index++) {
+    const napStartTime = napEndTime?.add(wakeWindows[index], 'hour') || null;
+    napEndTime = napStartTime?.add(napTime, 'hour') || null;
+    naps.push({
+      description: `Nap ${index + 1}`,
+      startTime: napStartTime?.toDate() || null,
+      endTime: napEndTime?.toDate() || null
+    });
+  }
+
+  const getReadyForBed = [
+    {
+      description: 'Get Ready for bed',
+      startTime: earliestGetReadyTime?.toDate()
+    }
+  ];
+
+  const asleep = [
+    {
+      description: 'Asleep ',
+      startTime: desiredBedTime?.toDate()
+    }
+  ];
+
+  const reminders = [...morningRise, ...naps, ...getReadyForBed, ...asleep];
+
+  const planData = {
+    clientId: userId,
+    coachId: user?.sub,
+    reminders: reminders
+  };
+
+  return planData;
+};
